I🏗️Integrated CMake buildsystem

+++ src/cxxnew.cpp
@@ -8,10 +8,10 @@
 
 #include <memory.hpp>
 
-void* operator new(unsigned long size) { return malloc(size); }
+void* operator new(size_t size) { return malloc(size); }
 void  operator delete(void* ptr) { free(ptr); }
-void  operator delete(void* ptr, unsigned long size) { free(ptr); }
-void* operator new[](unsigned long size) { return malloc(size); }
+void  operator delete(void* ptr, size_t size) { free(ptr); }
+void* operator new[](size_t size) { return malloc(size); }
 void  operator delete[](void* ptr) { free(ptr); }
 
 void *__dso_handle;
